backend/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
gunicorn==19.7.1
psycopg2==2.7.3.2
pytz==2017.3
supervisor==4.0.4